[INTERNAL] Set issue history default tab to "history" (trac/ 16409)
Changed issue history default tab from "notes" to "history".
Made "history" the default choice of the corresponding select field, if
no value was set so far.

As this setting wasn't present before redmine 4.1, so old users don't
have any value set for this setting. When saving personal preferences
Redmine defaults to the "notes" option. If no preference is set and
no matching cookie or history is present Redmine also defaulted to
showing the "notes" tab.

Since users complained about this not being the desired behavior, this
change was implemented.