[BUGFIX] Duplicates in user-selects (/trac 15608)
This change has been copied from the official Redmine 5.0.3. codebase:
Due to an error in checking for missing values, users that where
filteres for in queries where misidentified as missing and thus
added a second time to some selects. This was due to non-standard
data structure with the identifying bit of information not in the last
field of an array. With this change, the correct piece of information
is checked against.